feat(db,security): introduce Alembic migrations, Decimal-safe money/volume columns, and password self-service
Addresses the top four items from the build audit's priority-ranked
recommendation list (data integrity, then security):

- Alembic is now real: app/database/migrations.py wraps it
  programmatically, init_db() runs upgrade_to_head() instead of a blunt
  create_all(), and a pre-migration backup (app/database/backup.py, via
  sqlite3's online backup API so WAL-mode writes are never torn) is
  taken automatically whenever a launch detects pending migrations
  against a database that already has data in it.

- Every money/volume column (Fuel.rate_per_liter, Tank capacity/stock,
  TankReading, TankTransaction, FuelReconciliation, NozzleAssignment
  meters) moved from Float to Numeric/Decimal, closing the rounding-drift
  gap the audit flagged - all matching Pydantic schemas now type these
  fields as Decimal too, so the whole path from UI input to stored value
  is exact-decimal, not binary-float.

- Resolved the Fuel/Tank redundancy noted in the audit and in
  PROJECT_CONTEXT.md's known limitations: removed Fuel.capacity/
  current_stock/opening_stock (Tank already owns this data since Phase
  9) along with the dead, unused InventoryService that was the only
  thing still reading them.

- Found and fixed a real latent bug while testing the above:
  TankTransactionRepository.sum_for_tank_by_type compared a naive
  local-calendar-date boundary directly against transaction_at (stored
  UTC-aware), silently dropping transactions recorded after local
  midnight whenever local time runs ahead of UTC (e.g. IST, UTC+5:30) -
  fuel reconciliation could under-count receipts/issues for the current
  day. Fixed by converting the day-boundary to a UTC instant before
  comparing; pinned with two new deterministic regression tests.

- Password self-service: User.must_change_password forces a rotation
  after any admin-set password (new account creation or an explicit
  reset - UserService.reset_password is also new, closing the "a
  forgotten password has no recovery path" gap). ChangePasswordDialog is
  shown as an un-skippable step right after login when set, and is also
  reachable any time via a new "Change Password" button. The seeded
  default admin account is flagged too, so the known Admin@123 credential
  gets rotated on first real use rather than staying in place silently.

221/221 tests passing project-wide (17 new).

+++ app/models/fuel_reconciliation.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime, timezone
 
-from sqlalchemy import Column, Date, DateTime, Float, ForeignKey, String, Text
+from sqlalchemy import Column, Date, DateTime, ForeignKey, Numeric, String, Text
 from sqlalchemy.orm import relationship
 
 from .base import Base
@@ -23,13 +23,13 @@ class FuelReconciliation(Base):
     tank_id = Column(String(36), ForeignKey("tanks.id"), nullable=False, index=True)
     reconciliation_date = Column(Date, nullable=False, index=True)
 
-    opening_stock = Column(Float, nullable=False)
-    received_quantity = Column(Float, nullable=False)
-    sold_quantity = Column(Float, nullable=False)
-    expected_closing_stock = Column(Float, nullable=False)
-    physical_stock = Column(Float, nullable=False)
-    variance = Column(Float, nullable=False)
-    variance_percent = Column(Float, nullable=False)
+    opening_stock = Column(Numeric(12, 3), nullable=False)
+    received_quantity = Column(Numeric(12, 3), nullable=False)
+    sold_quantity = Column(Numeric(12, 3), nullable=False)
+    expected_closing_stock = Column(Numeric(12, 3), nullable=False)
+    physical_stock = Column(Numeric(12, 3), nullable=False)
+    variance = Column(Numeric(12, 3), nullable=False)
+    variance_percent = Column(Numeric(8, 3), nullable=False)
     classification = Column(String(32), nullable=False)
 
     performed_by_id = Column(String(36), ForeignKey("users.id"), nullable=False)
